Puts addPill modal in separate file to be reused
